% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordionPanelSelect.R
\name{accordionPanelSelect}
\alias{accordionPanelSelect}
\title{accordionPanelSelect}
\usage{
accordionPanelSelect(input, session, panel_id, accordion_id, menu_list)
}
\arguments{
\item{input}{The server input}

\item{session}{The server session}

\item{panel_id}{The shiny app ID for the nav panels.}

\item{accordion_id}{The shiny app ID for the accordion menu.}

\item{menu_list}{The nested list with menu and sub-menu items.}
}
\value{
An accordion list for domain and sub-domain
}
\description{
accordionPanelSelect
}
\author{
Braeden Klaver
}
