% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordionTabset.R
\name{accordionTabset}
\alias{accordionTabset}
\title{accordionTabset}
\usage{
accordionTabset(id, menu_list)
}
\value{
An accordion list for domain and sub-domain
}
\description{
accordionTabset
}
\author{
Braeden Klaver
}
