% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordionTabset.R
\name{accordionTabset}
\alias{accordionTabset}
\title{accordionTabset}
\usage{
accordionTabset(id, menu_list, class = NULL)
}
\arguments{
\item{id}{The shiny ID}

\item{menu_list}{The nested list with menu items and sub-items.}

\item{class}{The custom CSS class for the sub-item menu items}
}
\value{
An accordion list for domain and sub-domain
}
\description{
accordionTabset() creates the drop down navigation menu UI.
}
\author{
Braeden Klaver
}
